Flutter/Proyecto_Flutter_Optimizado/2Flutter/Flutter/Newclothesure/Backend/websocket/requirements_websocket.txt
# Dependencias para el servidor WebSocket
python-socketio==5.11.2
aiohttp==3.9.5
